lib/stream: Add and handle new VIDEO_SOURCE_ENCODED type
Add functions to handle a new entry into video_source_type, representing
data with an encoding step between the source and sink. When running
through this route, buffers are allocated and mmaped on both the source
and the sink, then imported to the source to be filled before being queued
to the sink.

Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>