uvc-gadget v0.2.0
This release extends UVC Gadget with the ability to provide a
'slideshow' source.

This can assist in the testing of performance and capabilities of a UVC
Gadget based device, by providing a known set of data frames and
configurations to stream to the host.

Further more this release contains:
 - lib/timer: Add timer infrastructure
 - lib/jpg-source: Constrain fps

These allow sources with out a controlling device to be configured to
deliver at the rate requested by the host.

Finally, the Makefile 'helper' is fixed to correctly call the install
target in ninja.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>