uvc-gadget v0.3.0
This release brings in some new support and features for UVC Gadget to
operate with libcamera based cameras.

 - libcamera sources are now supported
 - Support for encoded streams
 - Software based MJPEG encoding is managed by the libcamera source only.

Please be aware, that as part of this release, the existing V4L2 source
option '-c' is renamed to '-d'.

The new libcamera source is now supported by the option '-c' and can be
selected by camera index, or explictly choosing the camera based on it's
libcamera id.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>